Remove style that was lengthening certain cards
It was causing weird display when there are 7 cards, where it looked like the 7th card was jamming up into the 4th (because the 5th and 6th were longer). Removing it makes it so that the 5th and 6th cards are the same height as the 4th. Removing this style rule doesn't seem to have an adverse impact.